+++ src/include/duckdb/parser/parsed_data/attach_info.hpp
@@ -12,6 +12,7 @@
 #include "duckdb/common/vector.hpp"
 #include "duckdb/common/unordered_map.hpp"
 #include "duckdb/common/types/value.hpp"
+#include "duckdb/common/enums/on_create_conflict.hpp"
 
 namespace duckdb {
 
@@ -29,6 +30,8 @@ public:
 	string path;
 	//! Set of (key, value) options
 	unordered_map<string, Value> options;
+	//! What to do on create conflict
+	OnCreateConflict on_conflict = OnCreateConflict::ERROR_ON_CONFLICT;
 
 public:
 	unique_ptr<AttachInfo> Copy() const;
