+++ src/include/duckdb/storage/serialization/parse_info.json
@@ -287,6 +287,12 @@
         "id": 202,
         "name": "options",
         "type": "unordered_map<string, Value>"
+      },
+      {
+        "id": 203,
+        "name": "on_conflict",
+        "type": "OnCreateConflict",
+        "default": "OnCreateConflict::ERROR_ON_CONFLICT"
       }
     ]
   },
