+++ test/sql/attach/attach_if_not_exists.test
@@ -1,0 +1,19 @@
+# name: test/sql/attach/attach_if_not_exists.test
+# description: Test attaching of a read-only database
+# group: [attach]
+
+require skip_reload
+
+statement ok
+PRAGMA enable_verification
+
+statement ok
+ATTACH '__TEST_DIR__/attach_if_not_exists.db' AS db1
+
+statement ok
+ATTACH IF NOT EXISTS '__TEST_DIR__/attach_if_not_exists.db' AS db1
+
+statement error
+ATTACH '__TEST_DIR__/attach_if_not_exists.db' AS db1
+----
+already attached
